Make BukkitIntake less of a pain to use
I began testing the revised BukkitIntake system and it was a massive pain to use due
to the fact that users had to know exactly what each part of the system was responsible
for and instantiate everything in the correct order. So, I made the following changes
with the goal of leaving advanced features in place for users that need different
functionality and don't mind putting in extra work to learn how:
   - Removed the BukkitIntake builder
   - Moved parameters previously located in the BukkitIntake class elsewhere
      - Builder, Authorizer, and modules all moved to the CommandGraph
      - Injector moved to the Builder class
   - Added a BasicBukkitCommandGraph class that extends CommandGraph to make instantiating
     parameters very simple for people who don't want to add custom functionality

-- Personal Note --
I believe these changes make BukkitIntake more user friendly than ever before, however I
want to run more tests to make sure everything is working properly. Then, I'll send a PR.
(Yes, I realize this shouldn't be in the commit message but I'm keeping it)
 - ------------- -